+++ Decoder.h
@@ -7,12 +7,12 @@
 class HuffmanDecoder {
     private:
         Node* root;
-        int freq[256];
-        int tot_len;
+        unsigned int freq[256];
+        unsigned int tot_len;
         std::vector<uint8_t> compressed;
-        std::set <std::pair<int, Node*>> s;
+        std::set <std::pair<unsigned int, Node*>> s;
 
-        void add(int freq, int symbol, Node* left = nullptr, Node* right = nullptr);
+        void add(unsigned int freq, uint8_t symbol, Node* left = nullptr, Node* right = nullptr);
         void remTree(Node* curr);
         void buildTree();
         void readData();
